Create reusable view UnavailableAnalyticsView
